Move local H5 tests into optimized suite
