--- tests/integration/local_h5/__init__.py
@@ -1,1 +1,0 @@
-"""Local H5 integration tests."""
