+++ tests/optimized/local_h5/__init__.py
@@ -1,0 +1,1 @@
+"""Optimized local H5 integration tests."""
